tools/python/examples/openai/customer_support/pyproject.toml
[project]
name = "email-agent"
version = "0.1.0"
description = "An automated email support agent that uses AI to respond to customer support emails"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.0",
    "imaplib2==3.6",
    "python-decouple==3.8",
    "openai-agents==0.0.2",
    "stripe-agent-toolkit>=0.6.0",
    "stripe>=7.0.0",
    "urllib3<2.0.0",
    "markdown==3.7"
]

[tool.hatch.metadata]
allow-direct-references = true
